requirements.txt
#torch>=2.0.0
#pip3 install torch torchvision --index-url https://download.pytorch.org/whl/cu130
#triton>=3.0.0
cupy-cuda12x>=13.0.0
numpy>=1.24.0
pyyaml>=6.0
pytest>=7.0.0
#pip install cuda-tile

